+++ levior/__init__.py
@@ -2,7 +2,7 @@ from importlib import resources
 
 
 __appname__ = 'levior'
-__version__ = '1.3.1'
+__version__ = '1.3.2'
 
 
 def default_cert_paths() -> tuple: