+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.3.2] - 2024-03-04
+
+### Added
+
+- Proxy access restriction by IP address/subnet
+- Persist the access log in the diskcache (optional)
+
 ## [1.3.1] - 2024-03-03
 
 ### Added
@@ -6,6 +13,7 @@
 - Access log server endpoint
 - Feeds aggregator: make it possible to only show entries from a specific
   feed by passing its index in the URL query
+- Proxy access restriction by using IP address ACLs
 
 ## [1.3.0] - 2024-02-26